requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
matplotlib>=3.5.0
numpy>=1.21.0
Pillow>=8.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0
tensorboard>=2.8.0
seaborn>=0.11.0
albumentations>=1.3.0
segmentation-models-pytorch>=0.3.0
timm>=0.9.0
jupyter>=1.0.0
ipykernel>=6.0.0 